+++ src/jaegertracing/utils/UDPClient.h
@@ -60,7 +60,7 @@ class UDPClient : public agent::thrift::AgentIf {
             throw std::logic_error(oss.str());
         }
         const auto numWritten = ::write(_socket.handle(), data, size);
-        if (numWritten != size) {
+        if (static_cast<unsigned>(numWritten) != size) {
             std::ostringstream oss;
             oss << "Failed to write message"
                    ", numWritten="
@@ -75,7 +75,7 @@ class UDPClient : public agent::thrift::AgentIf {
 
   private:
     int _maxPacketSize;
-    boost::shared_ptr<apache::thrift::transport::TMemoryBuffer> _buffer;
+    std::shared_ptr<apache::thrift::transport::TMemoryBuffer> _buffer;
     net::Socket _socket;
     net::IPAddress _serverAddr;
     std::unique_ptr<agent::thrift::AgentClient> _client;